+++ Liturfaliar Cest/src/de/dakror/liturfaliar/item/skillanim/Sword0.java
@@ -113,7 +113,7 @@ public class Sword0 extends SkillAnimation
     AffineTransform t = AffineTransform.getRotateInstance(Math.toRadians(theta), caster.getRelativePos().x + rx + m.getX() + image.getWidth(null), caster.getRelativePos().y + ry + m.getY() + image.getHeight(null));
     
     g.setTransform(t);
-    g.drawImage(image, caster.getRelativePos().x + rx + m.getX(), caster.getRelativePos().y + ry + m.getY(), v.w);
+    g.drawImage(image, (int) caster.getRelativePos().x + rx + m.getX(), (int) caster.getRelativePos().y + ry + m.getY(), v.w);
     g.setTransform(oldTransform);
   }
